moved scripts in error dir
