+++ dictos/calculus/exceptions.py
