+++ dictos/discrete/exceptions.py
@@ -2,7 +2,7 @@
 Custom exceptions for errors related to stencil.
 """
 
-from .utils import find_duplicated_points
+from dictos.utilities.exceptions.utils import find_duplicated_points
 
 
 class StencilError(Exception):
