+++ dictos/linalg/exceptions.py
