+++ dictos/poly/exceptions.py
