+++ dictos/poly/interpolation.py
@@ -18,7 +18,7 @@ from dictos.utilities.utils import (
 from dictos.linalg.linalg import dot_product
 from dictos.poly.lagrangian_polynomial import lagrangian_poly
 from dictos.series.taylor_expansion import taylor_series
-from dictos.error.stencil import ContainsZeroError
+from dictos.discrete.exceptions import ContainsZeroError
 from dictos.core.expr import Expr
