+++ dictos/series/exceptions.py
