+++ dictos/utilities/exceptions/internal.py
