+++ dictos/utilities/exceptions/utils.py
